Move data.DataObs -> client.DataObsClient
