+++ cartoframes/client/__init__.py
@@ -1,7 +1,9 @@
 from __future__ import absolute_import
 
 from .sql_client import SQLClient
+from .data_obs_client import DataObsClient
 
 __all__ = [
-    'SQLClient'
+    'SQLClient',
+    'DataObsClient'
 ]
