+++ cartoframes/data/utils.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import re
 import os
 import sys
